refactor: 파일 경로 수정, 필요 없는 save 호출 제거
